Created the FruitDB class, which contains a field Map<String, Integer> fruitDataBase serving as a database.
To access FruitDB, the FruitDao interface and its implementation FruitDaoImpl were created.

Two custom exceptions were introduced: InvalidFruitDataException, indicating incorrect data input, and InvalidOperationException, thrown when attempting an undefined operation.

A FruitTransaction model was established to store data on fruit operations and details about the fruit.

For file reading, the ReadService interface and CsvReadServiceImpl implementation were developed.

To convert data from the file to JavaObject (FruitTransaction), the DataParserService interface and its implementation DataParserServiceImpl were created.

After conversion, we invoke the FruitOperationStrategy class and its getOperationService() method, which returns the necessary strategy based on the operation code.

Once all operations from the list of FruitTransaction are executed, we interact with the ReportService interface and its implementation ReportServiceImpl. The createReport() method returns a List<String> in the format "fruit,quantity."

Finally, we call WriteServiceImpl to write information to a CSV file.